+++ codegen/projections/rpcv2_cbor/lib/rpcv2_cbor/client.rb
@@ -9,11 +9,15 @@
 
 require 'stringio'
 
+require_relative 'plugins/global_config'
+
 module Rpcv2Cbor
   class Client < Hearth::Client
 
     # @api private
-    @plugins = Hearth::PluginList.new
+    @plugins = Hearth::PluginList.new([
+      Plugins::GlobalConfig.new
+    ])
 
     # @param [Hash] options
     #   Options used to construct an instance of {Config}