+++ codegen/projections/rpcv2_cbor/spec/protocol_spec.rb
@@ -40,9 +40,7 @@ module Rpcv2Cbor
           end
           interceptor = Hearth::Interceptor.new(read_before_transmit: proc)
           opts = {interceptors: [interceptor]}
-          client.empty_input_output({
-
-          }, **opts)
+          client.empty_input_output({}, **opts)
         end
 
       end
@@ -60,9 +58,7 @@ module Rpcv2Cbor
           client.stub_responses(:empty_input_output, response)
           allow(Builders::EmptyInputOutput).to receive(:build)
           output = client.empty_input_output({}, auth_resolver: Hearth::AnonymousAuthResolver.new)
-          expect(output.data.to_h).to match_cbor({
-
-          })
+          expect(output.data.to_h).to match_cbor({})
         end
 
         # When output structure is empty the client should accept an empty body
@@ -76,9 +72,7 @@ module Rpcv2Cbor
           client.stub_responses(:empty_input_output, response)
           allow(Builders::EmptyInputOutput).to receive(:build)
           output = client.empty_input_output({}, auth_resolver: Hearth::AnonymousAuthResolver.new)
-          expect(output.data.to_h).to match_cbor({
-
-          })
+          expect(output.data.to_h).to match_cbor({})
         end
 
       end
@@ -92,13 +86,9 @@ module Rpcv2Cbor
           end
           interceptor = Hearth::Interceptor.new(read_after_transmit: proc)
           allow(Builders::EmptyInputOutput).to receive(:build)
-          client.stub_responses(:empty_input_output, data: {
-
-          })
+          client.stub_responses(:empty_input_output, data: {})
           output = client.empty_input_output({}, interceptors: [interceptor], auth_resolver: Hearth::AnonymousAuthResolver.new)
-          expect(output.data.to_h).to match_cbor({
-
-          })
+          expect(output.data.to_h).to match_cbor({})
         end
 
         # When output structure is empty the client should accept an empty body
@@ -108,13 +98,9 @@ module Rpcv2Cbor
           end
           interceptor = Hearth::Interceptor.new(read_after_transmit: proc)
           allow(Builders::EmptyInputOutput).to receive(:build)
-          client.stub_responses(:empty_input_output, data: {
-
-          })
+          client.stub_responses(:empty_input_output, data: {})
           output = client.empty_input_output({}, interceptors: [interceptor], auth_resolver: Hearth::AnonymousAuthResolver.new)
-          expect(output.data.to_h).to match_cbor({
-
-          })
+          expect(output.data.to_h).to match_cbor({})
         end
 
       end
@@ -436,25 +422,19 @@ module Rpcv2Cbor
           begin
             output = client.greeting_with_errors({}, auth_resolver: Hearth::AnonymousAuthResolver.new)
           rescue Errors::ComplexError => e
-            expect(e.data.to_h).to eq({
-
-            })
+            expect(e.data.to_h).to eq({})
           end
         end
 
         #
         it 'stubs RpcV2CborEmptyComplexError' do
-          client.stub_responses(:greeting_with_errors, error: { class: Errors::ComplexError, data: {
-
-          } })
+          client.stub_responses(:greeting_with_errors, error: { class: Errors::ComplexError, data: {} })
           allow(Builders::GreetingWithErrors).to receive(:build)
           begin
             output = client.greeting_with_errors({}, auth_resolver: Hearth::AnonymousAuthResolver.new)
           rescue Errors::ComplexError => e
             expect(e.http_status).to eq(400)
-            expect(e.data.to_h).to eq({
-
-            })
+            expect(e.data.to_h).to eq({})
           end
         end
       end
@@ -477,9 +457,7 @@ module Rpcv2Cbor
           end
           interceptor = Hearth::Interceptor.new(read_before_transmit: proc)
           opts = {interceptors: [interceptor]}
-          client.no_input_output({
-
-          }, **opts)
+          client.no_input_output({}, **opts)
         end
 
       end
@@ -496,9 +474,7 @@ module Rpcv2Cbor
           client.stub_responses(:no_input_output, response)
           allow(Builders::NoInputOutput).to receive(:build)
           output = client.no_input_output({}, auth_resolver: Hearth::AnonymousAuthResolver.new)
-          expect(output.data.to_h).to match_cbor({
-
-          })
+          expect(output.data.to_h).to match_cbor({})
         end
 
         # Clients should accept a CBOR empty struct if there is no output.
@@ -512,9 +488,7 @@ module Rpcv2Cbor
           client.stub_responses(:no_input_output, response)
           allow(Builders::NoInputOutput).to receive(:build)
           output = client.no_input_output({}, auth_resolver: Hearth::AnonymousAuthResolver.new)
-          expect(output.data.to_h).to match_cbor({
-
-          })
+          expect(output.data.to_h).to match_cbor({})
         end
 
         # Clients should accept an empty body if there is no output and
@@ -529,9 +503,7 @@ module Rpcv2Cbor
           client.stub_responses(:no_input_output, response)
           allow(Builders::NoInputOutput).to receive(:build)
           output = client.no_input_output({}, auth_resolver: Hearth::AnonymousAuthResolver.new)
-          expect(output.data.to_h).to match_cbor({
-
-          })
+          expect(output.data.to_h).to match_cbor({})
         end
 
       end
@@ -545,13 +517,9 @@ module Rpcv2Cbor
           end
           interceptor = Hearth::Interceptor.new(read_after_transmit: proc)
           allow(Builders::NoInputOutput).to receive(:build)
-          client.stub_responses(:no_input_output, data: {
-
-          })
+          client.stub_responses(:no_input_output, data: {})
           output = client.no_input_output({}, interceptors: [interceptor], auth_resolver: Hearth::AnonymousAuthResolver.new)
-          expect(output.data.to_h).to match_cbor({
-
-          })
+          expect(output.data.to_h).to match_cbor({})
         end
 
         # Clients should accept a CBOR empty struct if there is no output.
@@ -561,13 +529,9 @@ module Rpcv2Cbor
           end
           interceptor = Hearth::Interceptor.new(read_after_transmit: proc)
           allow(Builders::NoInputOutput).to receive(:build)
-          client.stub_responses(:no_input_output, data: {
-
-          })
+          client.stub_responses(:no_input_output, data: {})
           output = client.no_input_output({}, interceptors: [interceptor], auth_resolver: Hearth::AnonymousAuthResolver.new)
-          expect(output.data.to_h).to match_cbor({
-
-          })
+          expect(output.data.to_h).to match_cbor({})
         end
 
         # Clients should accept an empty body if there is no output and
@@ -578,13 +542,9 @@ module Rpcv2Cbor
           end
           interceptor = Hearth::Interceptor.new(read_after_transmit: proc)
           allow(Builders::NoInputOutput).to receive(:build)
-          client.stub_responses(:no_input_output, data: {
-
-          })
+          client.stub_responses(:no_input_output, data: {})
           output = client.no_input_output({}, interceptors: [interceptor], auth_resolver: Hearth::AnonymousAuthResolver.new)
-          expect(output.data.to_h).to match_cbor({
-
-          })
+          expect(output.data.to_h).to match_cbor({})
         end
 
       end
@@ -608,9 +568,7 @@ module Rpcv2Cbor
           interceptor = Hearth::Interceptor.new(read_before_transmit: proc)
           opts = {interceptors: [interceptor]}
           client.operation_with_defaults({
-            defaults: {
-
-            }
+            defaults: {}
           }, **opts)
         end
 
@@ -626,9 +584,7 @@ module Rpcv2Cbor
           end
           interceptor = Hearth::Interceptor.new(read_before_transmit: proc)
           opts = {interceptors: [interceptor]}
-          client.operation_with_defaults({
-
-          }, **opts)
+          client.operation_with_defaults({}, **opts)
         end
 
         # Client uses explicitly provided member values over defaults
@@ -707,9 +663,7 @@ module Rpcv2Cbor
           interceptor = Hearth::Interceptor.new(read_before_transmit: proc)
           opts = {interceptors: [interceptor]}
           client.operation_with_defaults({
-            client_optional_defaults: {
-
-            }
+            client_optional_defaults: {}
           }, **opts)
         end
 
@@ -961,9 +915,7 @@ module Rpcv2Cbor
           end
           interceptor = Hearth::Interceptor.new(read_before_transmit: proc)
           opts = {interceptors: [interceptor]}
-          client.optional_input_output({
-
-          }, **opts)
+          client.optional_input_output({}, **opts)
         end
 
       end
@@ -981,9 +933,7 @@ module Rpcv2Cbor
           client.stub_responses(:optional_input_output, response)
           allow(Builders::OptionalInputOutput).to receive(:build)
           output = client.optional_input_output({}, auth_resolver: Hearth::AnonymousAuthResolver.new)
-          expect(output.data.to_h).to match_cbor({
-
-          })
+          expect(output.data.to_h).to match_cbor({})
         end
 
       end
@@ -997,13 +947,9 @@ module Rpcv2Cbor
           end
           interceptor = Hearth::Interceptor.new(read_after_transmit: proc)
           allow(Builders::OptionalInputOutput).to receive(:build)
-          client.stub_responses(:optional_input_output, data: {
-
-          })
+          client.stub_responses(:optional_input_output, data: {})
           output = client.optional_input_output({}, interceptors: [interceptor], auth_resolver: Hearth::AnonymousAuthResolver.new)
-          expect(output.data.to_h).to match_cbor({
-
-          })
+          expect(output.data.to_h).to match_cbor({})
         end
 
       end
@@ -1351,8 +1297,7 @@ module Rpcv2Cbor
               'y' => [
                 "a",
                 "b"
-              ],
-              'z' => nil
+              ]
             }
           })
         end
@@ -1465,8 +1410,7 @@ module Rpcv2Cbor
               'y' => [
                 "a",
                 "b"
-              ],
-              'z' => nil
+              ]
             }
           })
           output = client.rpc_v2_cbor_dense_maps({}, interceptors: [interceptor], auth_resolver: Hearth::AnonymousAuthResolver.new)
@@ -1478,8 +1422,7 @@ module Rpcv2Cbor
               'y' => [
                 "a",
                 "b"
-              ],
-              'z' => nil
+              ]
             }
           })
         end
@@ -2354,7 +2297,7 @@ module Rpcv2Cbor
             expect(request.uri.path).to eq('/service/RpcV2Protocol/operation/SimpleScalarProperties')
             { 'Content-Type' => 'application/cbor', 'smithy-protocol' => 'rpc-v2-cbor' }.each { |k, v| expect(request.headers[k]).to eq(v) }
             ['Content-Length'].each { |k| expect(request.headers.key?(k)).to be(true) }
-            expect(Hearth::CBOR.decode(request.body.read)).to match_cbor(Hearth::CBOR.decode(::Base64.decode64('v2lieXRlVmFsdWUFa2RvdWJsZVZhbHVl+z/+OVgQYk3TcWZhbHNlQm9vbGVhblZhbHVl9GpmbG9hdFZhbHVl+kDz989saW50ZWdlclZhbHVlGQEAaWxvbmdWYWx1ZRkmkWpzaG9ydFZhbHVlGSaqa3N0cmluZ1ZhbHVlZnNpbXBsZXB0cnVlQm9vbGVhblZhbHVl9WlibG9iVmFsdWVDZm9v/w==')))
+            expect(Hearth::CBOR.decode(request.body.read)).to match_cbor(Hearth::CBOR.decode(::Base64.decode64('v2lieXRlVmFsdWUFa2RvdWJsZVZhbHVl+z/+OVgQYk3TcWZhbHNlQm9vbGVhblZhbHVl9GpmbG9hdFZhbHVl+kD0AABsaW50ZWdlclZhbHVlGQEAaWxvbmdWYWx1ZRkmkWpzaG9ydFZhbHVlGSaqa3N0cmluZ1ZhbHVlZnNpbXBsZXB0cnVlQm9vbGVhblZhbHVl9WlibG9iVmFsdWVDZm9v/w==')))
           end
           interceptor = Hearth::Interceptor.new(read_before_transmit: proc)
           opts = {interceptors: [interceptor]}
@@ -2362,7 +2305,7 @@ module Rpcv2Cbor
             byte_value: 5,
             double_value: 1.889,
             false_boolean_value: false,
-            float_value: 7.624,
+            float_value: 7.625,
             integer_value: 256,
             long_value: 9873,
             short_value: 9898,
@@ -2453,7 +2396,7 @@ module Rpcv2Cbor
           response.status = 200
           response.headers['Content-Type'] = 'application/cbor'
           response.headers['smithy-protocol'] = 'rpc-v2-cbor'
-          response.body.write(::Base64.decode64('v3B0cnVlQm9vbGVhblZhbHVl9XFmYWxzZUJvb2xlYW5WYWx1ZfRpYnl0ZVZhbHVlBWtkb3VibGVWYWx1Zfs//jlYEGJN02pmbG9hdFZhbHVl+kDz989saW50ZWdlclZhbHVlGQEAanNob3J0VmFsdWUZJqprc3RyaW5nVmFsdWVmc2ltcGxlaWJsb2JWYWx1ZUNmb2//'))
+          response.body.write(::Base64.decode64('v3B0cnVlQm9vbGVhblZhbHVl9XFmYWxzZUJvb2xlYW5WYWx1ZfRpYnl0ZVZhbHVlBWtkb3VibGVWYWx1Zfs//jlYEGJN02pmbG9hdFZhbHVl+kD0AABsaW50ZWdlclZhbHVlGQEAanNob3J0VmFsdWUZJqprc3RyaW5nVmFsdWVmc2ltcGxlaWJsb2JWYWx1ZUNmb2//'))
           response.body.rewind
           client.stub_responses(:simple_scalar_properties, response)
           allow(Builders::SimpleScalarProperties).to receive(:build)
@@ -2463,7 +2406,7 @@ module Rpcv2Cbor
             false_boolean_value: false,
             byte_value: 5,
             double_value: 1.889,
-            float_value: 7.624,
+            float_value: 7.625,
             integer_value: 256,
             short_value: 9898,
             string_value: "simple",
@@ -2477,7 +2420,7 @@ module Rpcv2Cbor
           response.status = 200
           response.headers['Content-Type'] = 'application/cbor'
           response.headers['smithy-protocol'] = 'rpc-v2-cbor'
-          response.body.write(::Base64.decode64('qXB0cnVlQm9vbGVhblZhbHVl9XFmYWxzZUJvb2xlYW5WYWx1ZfRpYnl0ZVZhbHVlBWtkb3VibGVWYWx1Zfs//jlYEGJN02pmbG9hdFZhbHVl+kDz989saW50ZWdlclZhbHVlGQEAanNob3J0VmFsdWUZJqprc3RyaW5nVmFsdWVmc2ltcGxlaWJsb2JWYWx1ZUNmb28='))
+          response.body.write(::Base64.decode64('qXB0cnVlQm9vbGVhblZhbHVl9XFmYWxzZUJvb2xlYW5WYWx1ZfRpYnl0ZVZhbHVlBWtkb3VibGVWYWx1Zfs//jlYEGJN02pmbG9hdFZhbHVl+kD0AABsaW50ZWdlclZhbHVlGQEAanNob3J0VmFsdWUZJqprc3RyaW5nVmFsdWVmc2ltcGxlaWJsb2JWYWx1ZUNmb28='))
           response.body.rewind
           client.stub_responses(:simple_scalar_properties, response)
           allow(Builders::SimpleScalarProperties).to receive(:build)
@@ -2487,7 +2430,7 @@ module Rpcv2Cbor
             false_boolean_value: false,
             byte_value: 5,
             double_value: 1.889,
-            float_value: 7.624,
+            float_value: 7.625,
             integer_value: 256,
             short_value: 9898,
             string_value: "simple",
@@ -2506,9 +2449,7 @@ module Rpcv2Cbor
           client.stub_responses(:simple_scalar_properties, response)
           allow(Builders::SimpleScalarProperties).to receive(:build)
           output = client.simple_scalar_properties({}, auth_resolver: Hearth::AnonymousAuthResolver.new)
-          expect(output.data.to_h).to match_cbor({
-
-          })
+          expect(output.data.to_h).to match_cbor({})
         end
 
         # Supports handling NaN float values.
@@ -2590,7 +2531,7 @@ module Rpcv2Cbor
           response.status = 200
           response.headers['Content-Type'] = 'application/cbor'
           response.headers['smithy-protocol'] = 'rpc-v2-cbor'
-          response.body.write(::Base64.decode64('v2lieXRlVmFsdWUFa2RvdWJsZVZhbHVl+z/+OVgQYk3TcWZhbHNlQm9vbGVhblZhbHVl9GpmbG9hdFZhbHVl+kDz989rZXh0cmFPYmplY3S/c2luZGVmaW5pdGVMZW5ndGhNYXC/a3dpdGhBbkFycmF5nwECA///cWRlZmluaXRlTGVuZ3RoTWFwo3J3aXRoQURlZmluaXRlQXJyYXmDAQIDeB1hbmRTb21lSW5kZWZpbml0ZUxlbmd0aFN0cmluZ3gfdGhhdCBoYXMsIGJlZW4gY2h1bmtlZCBvbiBjb21tYWxub3JtYWxTdHJpbmdjZm9vanNob3J0VmFsdWUZJw9uc29tZU90aGVyRmllbGR2dGhpcyBzaG91bGQgYmUgc2tpcHBlZP9saW50ZWdlclZhbHVlGQEAaWxvbmdWYWx1ZRkmkWpzaG9ydFZhbHVlGSaqa3N0cmluZ1ZhbHVlZnNpbXBsZXB0cnVlQm9vbGVhblZhbHVl9WlibG9iVmFsdWVDZm9v/w=='))
+          response.body.write(::Base64.decode64('v2lieXRlVmFsdWUFa2RvdWJsZVZhbHVl+z/+OVgQYk3TcWZhbHNlQm9vbGVhblZhbHVl9GpmbG9hdFZhbHVl+kD0AABrZXh0cmFPYmplY3S/c2luZGVmaW5pdGVMZW5ndGhNYXC/a3dpdGhBbkFycmF5nwECA///cWRlZmluaXRlTGVuZ3RoTWFwo3J3aXRoQURlZmluaXRlQXJyYXmDAQIDeB1hbmRTb21lSW5kZWZpbml0ZUxlbmd0aFN0cmluZ3gfdGhhdCBoYXMsIGJlZW4gY2h1bmtlZCBvbiBjb21tYWxub3JtYWxTdHJpbmdjZm9vanNob3J0VmFsdWUZJw9uc29tZU90aGVyRmllbGR2dGhpcyBzaG91bGQgYmUgc2tpcHBlZP9saW50ZWdlclZhbHVlGQEAaWxvbmdWYWx1ZRkmkWpzaG9ydFZhbHVlGSaqa3N0cmluZ1ZhbHVlZnNpbXBsZXB0cnVlQm9vbGVhblZhbHVl9WlibG9iVmFsdWVDZm9v/w=='))
           response.body.rewind
           client.stub_responses(:simple_scalar_properties, response)
           allow(Builders::SimpleScalarProperties).to receive(:build)
@@ -2599,7 +2540,7 @@ module Rpcv2Cbor
             byte_value: 5,
             double_value: 1.889,
             false_boolean_value: false,
-            float_value: 7.624,
+            float_value: 7.625,
             integer_value: 256,
             long_value: 9873,
             short_value: 9898,
@@ -2625,7 +2566,7 @@ module Rpcv2Cbor
             false_boolean_value: false,
             byte_value: 5,
             double_value: 1.889,
-            float_value: 7.624,
+            float_value: 7.625,
             integer_value: 256,
             short_value: 9898,
             string_value: "simple",
@@ -2637,7 +2578,7 @@ module Rpcv2Cbor
             false_boolean_value: false,
             byte_value: 5,
             double_value: 1.889,
-            float_value: 7.624,
+            float_value: 7.625,
             integer_value: 256,
             short_value: 9898,
             string_value: "simple",
@@ -2657,7 +2598,7 @@ module Rpcv2Cbor
             false_boolean_value: false,
             byte_value: 5,
             double_value: 1.889,
-            float_value: 7.624,
+            float_value: 7.625,
             integer_value: 256,
             short_value: 9898,
             string_value: "simple",
@@ -2669,7 +2610,7 @@ module Rpcv2Cbor
             false_boolean_value: false,
             byte_value: 5,
             double_value: 1.889,
-            float_value: 7.624,
+            float_value: 7.625,
             integer_value: 256,
             short_value: 9898,
             string_value: "simple",
@@ -2684,13 +2625,9 @@ module Rpcv2Cbor
           end
           interceptor = Hearth::Interceptor.new(read_after_transmit: proc)
           allow(Builders::SimpleScalarProperties).to receive(:build)
-          client.stub_responses(:simple_scalar_properties, data: {
-
-          })
+          client.stub_responses(:simple_scalar_properties, data: {})
           output = client.simple_scalar_properties({}, interceptors: [interceptor], auth_resolver: Hearth::AnonymousAuthResolver.new)
-          expect(output.data.to_h).to match_cbor({
-
-          })
+          expect(output.data.to_h).to match_cbor({})
         end
 
         # Supports handling NaN float values.
@@ -2784,7 +2721,7 @@ module Rpcv2Cbor
             byte_value: 5,
             double_value: 1.889,
             false_boolean_value: false,
-            float_value: 7.624,
+            float_value: 7.625,
             integer_value: 256,
             long_value: 9873,
             short_value: 9898,
@@ -2797,7 +2734,7 @@ module Rpcv2Cbor
             byte_value: 5,
             double_value: 1.889,
             false_boolean_value: false,
-            float_value: 7.624,
+            float_value: 7.625,
             integer_value: 256,
             long_value: 9873,
             short_value: 9898,