+++ codegen/smithy-ruby-codegen/src/main/java/software/amazon/smithy/ruby/codegen/DirectedRubyCodegen.java
@@ -43,6 +43,7 @@ import software.amazon.smithy.ruby.codegen.generators.ConfigGenerator;
 import software.amazon.smithy.ruby.codegen.generators.EndpointGenerator;
 import software.amazon.smithy.ruby.codegen.generators.EventStreamGenerator;
 import software.amazon.smithy.ruby.codegen.generators.GemspecGenerator;
+import software.amazon.smithy.ruby.codegen.generators.GlobalConfigPluginGenerator;
 import software.amazon.smithy.ruby.codegen.generators.HttpProtocolTestGenerator;
 import software.amazon.smithy.ruby.codegen.generators.MiddlewareGenerator;
 import software.amazon.smithy.ruby.codegen.generators.ModuleGenerator;
@@ -176,6 +177,9 @@ public class DirectedRubyCodegen
         ConfigGenerator configGenerator = new ConfigGenerator(directive, clientConfigList);
         configGenerator.render();
         configGenerator.renderRbs();
+
+        GlobalConfigPluginGenerator globalConfigPluginGenerator = new GlobalConfigPluginGenerator(directive);
+        globalConfigPluginGenerator.render();
 
         MiddlewareGenerator middlewareGenerator = new MiddlewareGenerator(directive, middlewareBuilder);
         middlewareGenerator.render();